+++ pkg/commands/service/logging/doc.go
@@ -1,3 +1,2 @@
-// Package logging contains commands to inspect and manipulate Fastly service
-// logging endpoints.
+// Package logging contains commands to inspect and manage Fastly service logging.
 package logging